Add support for inverting stop conditions
Stop conditions can now be inverted using the `~` operator.

This is useful for creating stop conditions that are the negation of other stop conditions.

Example:
Create a stop condition that stops retrying if the exception is not a `RuntimeError`.
```python
from mule.stop_conditions import ExceptionMatches,

stop_condition = ~ExceptionMatches(RuntimeError)
```